requirements.txt
requests
schedule
python-dotenv
feedparser
pandas
apscheduler